feat(node): address review — pin action, fix doc path, widen offset, guard cross step
Copilot review on #386:
- publish-node.yml: pin `dtolnay/rust-toolchain` to its commit SHA (CodeQL flagged
  the floating `@stable` tag for a third-party action).
- publish-node.yml: the aarch64 cross-linker step used `if: cross-linker != ''`,
  which is truthy for unset matrix entries on macOS/Windows — switched to the
  truthy `if: matrix.settings.cross-linker` so the apt step only runs on the
  Linux-arm64 entry.
- src/lib.rs: the module doc pointed at `src/index.ts`; the idiomatic layer is
  `index.ts` at the binding root. Fixed.
- src/lib.rs: `Untranslatable.offset` was `usize as u32`, which could truncate on
  a >4 GiB input; widened to `i64` (still a JS `number`), so no real byte offset
  truncates.

Re-verified: napi build + vitest 35/35, tsc --noEmit, cargo fmt/clippy clean.

+++ bindings/node/src/lib.rs
@@ -4,7 +4,7 @@
 //! **raw native shim**: each function is a thin wrapper over `disarm_core::api`
 //! with positional arguments and string-token enums. The idiomatic TypeScript
 //! surface — options objects, string-union types, sensible defaults, and the
-//! `DisarmError` class — lives in the hand-written `src/index.ts` layer that calls
+//! `DisarmError` class — lives in the hand-written `index.ts` layer that calls
 //! these. Keeping the shim dumb lets the core's surface be reshaped without
 //! re-teaching the binding its idioms.
 //!
@@ -81,7 +81,9 @@ pub fn find_untranslatable(
         .into_iter()
         .map(|u| Untranslatable {
             char: u.ch.to_string(),
-            offset: u.offset as u32,
+            // i64 (→ JS number) holds any real byte offset without the u32
+            // truncation a >4 GiB input could otherwise cause.
+            offset: u.offset as i64,
         })
         .collect())
 }
@@ -92,7 +94,7 @@ pub struct Untranslatable {
     /// The untranslatable character.
     pub char: String,
     /// Its byte offset in the input string.
-    pub offset: u32,
+    pub offset: i64,
 }
 
 // ── Confusables (TR39) ────────────────────────────────────────────────────────
